+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Usage Consolidation: Unable to handle gracefully handle when user attempts to view titles over 200,000. (UIEH-1136)
 * Add tests for ResourceEditManagedTitle component. (UIEH-1096)
 * Avoid .all permissions. (UIEH-1135)
+* Add tests for CoverageDateList component. (UIEH-1056)
 
 ## [6.1.0] (https://github.com/folio-org/ui-eholdings/tree/v6.1.0) (2021-06-04)
 * Add tests coverage for keyboard shortcuts to eholdings. (UIEH-1043)